Add Image/Flavor Constraint to Rackspace Server
The image and flavor used to build a nova server must satisfy
certain properties to successfully build the server.
Currently, we do not check for valid combinations before
invoking a call to nova. This patch adds support to the
Rackspace Cloud Server resource to enable such checking so
that invalid combinations are discovered before stack
creation begins.

Change-Id: I78a006b0004d28b04c7abc9a4e83d211c5437e82
Co-Authored-By: Anna Eilering <anna.eilering@rackspace.com>