Store the raw, unparsed template in heat
Currently, only a parsed version of the template is stored in the
database. As a result, any comments that the user may have supplied
are lost. Storing the unparsed version of the template will remedy
this by allowing us to show the original version of the template
used to create a stack in a call to 'get template'.

Change-Id: I2d6e317cd6fbfc83b1c83a73c1de905f699b763f
Implements: blueprint store-raw-template